+++ backend/api.py
@@ -1,0 +1,5 @@
+from ninja import NinjaAPI
+from produk.api import router as produk_router
+
+api = NinjaAPI()
+api.add_router("/produk", produk_router)
