+++ backend/settings.py
@@ -41,7 +41,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'api',
+    'produk'
 ]
 
 MIDDLEWARE = [
