+++ backend/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from api.views import api
+from backend.api import api
 
 urlpatterns = [
     path("admin/", admin.site.urls),
